renamed the class names and added two new classes for values, messages checks
