+++ tests/acceptance-tests/features/steps/bmi_survey_steps.py
@@ -1,7 +1,5 @@
 from behave import given, when, then
-
-from pages.common.contributor_details.check_contributor_details import CheckContributorDetails
-from pages.common.contributor_details.validate_contributor_details import ValidateContributorDetails
+from pages.common.contributor_details_page import ContributorDetailsPage
 
 
 @given(u'I run the validation process on {question_codes} with {period_value}')
@@ -11,7 +9,7 @@ from pages.common.contributor_details.validate_contributor_details import Valida
 def run_the_validation_process(context, period_value, question_codes=None):
     if not question_codes:
         question_codes = context.question_codes
-    page = ValidateContributorDetails()
+    page = ContributorDetailsPage()
     if context.period_type == "previous":
         context.comparison_value_one = period_value
         page.validate_the_previous_period_details(context.survey, question_codes, context.comparison_value_one)
@@ -24,7 +22,7 @@ def run_the_validation_process(context, period_value, question_codes=None):
 @then(
     u'the validation should return {result} if the "{validation_check}"')
 def check_validation(context, result, validation_check):
-    page = CheckContributorDetails()
+    page = ContributorDetailsPage()
     if validation_check == 'values are not equal':
         page.check_values_are_not_equal(context.question_codes, context.comparison_value_one,
                                         context.comparison_value_two, result)
