+++ tests/acceptance-tests/features/steps/cognito_steps.py
@@ -3,13 +3,15 @@ import json
 from behave import given
 from pages.common.login_page import LoginPage
 
+
 def user_modified(context, username, password, roles):
     if username in context.created_users:
         return not (
-            context.created_users[username]["password"] == password
-            and context.created_users[username]["roles"].sort() == roles.sort()
+                context.created_users[username]["password"] == password
+                and context.created_users[username]["roles"].sort() == roles.sort()
         )
     return True
+
 
 @given(u'"{username}" exists with the password "{password}" and roles {roles}')
 def create_user(context, username, password, roles):
@@ -22,6 +24,7 @@ def create_user(context, username, password, roles):
     else:
         context.created_users[username] = {"password": password, "roles": roles}
         context.cognito.create_user(username, password, roles)
+
 
 @given(u'"{username}" is logged in')
 def login(context, username):
